Moved name ownership actions to its own module
