+++ test/unit/src/common/actions/t_register-name.js
@@ -36,12 +36,6 @@ const {
   registerNameSuccess,
   registerNameError,
   registerNameClear,
-  internalNameOwnership,
-  checkNameOwnership,
-
-  NAME_OWNERSHIP_NOT_REGISTERED,
-  NAME_OWNERSHIP_ALREADY_OWNED,
-  NAME_OWNERSHIP_REGISTERED_BY_OTHER_USER
 } = registerNameModule;
 const ActionTypes = registerNameModule;
 
@@ -466,229 +460,6 @@ describe('register name actions', () => {
 
     it('creates a valid flux standard action', () => {
       expect(isFSA(action)).toBe(true);
-    });
-  });
-
-  context('internalNameOwnership', () => {
-
-    context('when given snap name is not registered in the store', () => {
-      let api;
-      const snapName = 'test-snap';
-
-      beforeEach(() => {
-        api = nock(conf.get('STORE_API_URL'))
-          .post('/acl/', {
-            packages: [{ name: snapName }],
-            permissions: ['package_upload']
-          })
-          .reply(400, { status: 'error' });
-      });
-
-      afterEach(() => {
-        api.done();
-        nock.cleanAll();
-      });
-
-      it('should return NAME_OWNERSHIP_NOT_REGISTERED status code', async () => {
-        const status = await internalNameOwnership(root, discharge, snapName);
-
-        expect(status).toBe(NAME_OWNERSHIP_NOT_REGISTERED);
-      });
-    });
-
-    context('when given snap name is registered by current user', () => {
-      let storeApi;
-      let bsiApi;
-
-      const snapName = 'test-snap';
-
-      beforeEach(() => {
-        storeApi = nock(conf.get('STORE_API_URL'))
-          .post('/acl/', {
-            packages: [{ name: snapName }],
-            permissions: ['package_upload']
-          })
-          .reply(200, { macaroon: 'test-macaroon' });
-        bsiApi = nock(BASE_URL)
-          .post('/api/store/register-name', { snap_name: snapName })
-          .reply(409, { code: 'already_owned' });
-      });
-
-      afterEach(() => {
-        storeApi.done();
-        bsiApi.done();
-        nock.cleanAll();
-      });
-
-      it('should return NAME_OWNERSHIP_ALREADY_OWNED status code', async () => {
-        const status = await internalNameOwnership(root, discharge, snapName);
-
-        expect(status).toBe(NAME_OWNERSHIP_ALREADY_OWNED);
-      });
-    });
-
-    context('when given snap name is registered by another user', () => {
-      let storeApi;
-      let bsiApi;
-
-      const snapName = 'test-snap';
-
-      beforeEach(() => {
-        storeApi = nock(conf.get('STORE_API_URL'))
-          .post('/acl/', {
-            packages: [{ name: snapName }],
-            permissions: ['package_upload']
-          })
-          .reply(200, { macaroon: 'test-macaroon' });
-        bsiApi = nock(BASE_URL)
-          .post('/api/store/register-name', { snap_name: snapName })
-          .reply(409, { code: 'already_registered' });
-      });
-
-      afterEach(() => {
-        storeApi.done();
-        bsiApi.done();
-        nock.cleanAll();
-      });
-
-      it('should return NAME_OWNERSHIP_REGISTERED_BY_OTHER_USER status code', async () => {
-        const status = await internalNameOwnership(root, discharge, snapName);
-
-        expect(status).toBe(NAME_OWNERSHIP_REGISTERED_BY_OTHER_USER);
-      });
-    });
-
-    context('when given snap name is registered by another user', () => {
-      let storeApi;
-      let bsiApi;
-
-      const snapName = 'test-snap';
-
-      beforeEach(() => {
-        storeApi = nock(conf.get('STORE_API_URL'))
-          .post('/acl/', {
-            packages: [{ name: snapName }],
-            permissions: ['package_upload']
-          })
-          .reply(200, { macaroon: 'test-macaroon' });
-        bsiApi = nock(BASE_URL)
-          .post('/api/store/register-name', { snap_name: snapName })
-          .reply(200, { status: 'success' });
-      });
-
-      afterEach(() => {
-        storeApi.done();
-        bsiApi.done();
-        nock.cleanAll();
-      });
-
-      it('should throw an error with "unexpected-register-response" code', async () => {
-        try {
-          await internalNameOwnership(root, discharge, snapName);
-          throw new Error('Unexpected success.');
-        } catch (error) {
-          expect(error.json).toInclude({
-            code: 'unexpected-register-response'
-          });
-        }
-      });
-    });
-  });
-
-  context('checkNameOwnership', () => {
-    const name = 'test-snap';
-    let api;
-
-
-    beforeEach(() => {
-      localForageStub.store['package_upload_request'] = { root, discharge };
-    });
-
-    context('when snapName is not specified', () => {
-      it('should throw an error', async () => {
-        try {
-          await store.dispatch(checkNameOwnership(''));
-          throw new Error('Unexpected success');
-        } catch (error) {
-          expect(error.message).toContain('`name` is required param');
-        }
-      });
-    });
-
-    context('when name ownership successfully verified', () => {
-      beforeEach(() => {
-        api = nock(conf.get('STORE_API_URL'))
-          .post('/acl/', {
-            packages: [{ name }],
-            permissions: ['package_upload']
-          })
-          .reply(404, {
-            status: 404,
-            error_code: 'resource-not-found'
-          });
-      });
-
-      afterEach(() => {
-        api.done();
-        nock.cleanAll();
-      });
-
-      it('should store CHECK_NAME_OWNERSHIP_REQUEST action', async() => {
-        const expectedAction = {
-          type: ActionTypes.CHECK_NAME_OWNERSHIP_REQUEST,
-          payload: { name }
-        };
-        await store.dispatch(checkNameOwnership(name));
-        expect(store.getActions()).toInclude(expectedAction);
-      });
-
-
-      it('should store CHECK_NAME_OWNERSHIP_SUCCESS action', async() => {
-        const expectedAction = {
-          type: ActionTypes.CHECK_NAME_OWNERSHIP_SUCCESS,
-          payload: {
-            name,
-            status: NAME_OWNERSHIP_NOT_REGISTERED
-          }
-        };
-        await store.dispatch(checkNameOwnership(name));
-        expect(store.getActions()).toInclude(expectedAction);
-      });
-
-    });
-
-    context('when name ownership thrown an error', () => {
-      beforeEach(() => {
-        api = nock(conf.get('STORE_API_URL'))
-          .post('/acl/', {
-            packages: [{ name }],
-            permissions: ['package_upload']
-          })
-          .reply(500, '<html>ERROR!</html>');
-      });
-
-      afterEach(() => {
-        api.done();
-        nock.cleanAll();
-      });
-
-      it('should store CHECK_NAME_OWNERSHIP_REQUEST action', async() => {
-        const expectedAction = {
-          type: ActionTypes.CHECK_NAME_OWNERSHIP_REQUEST,
-          payload: { name }
-        };
-        await store.dispatch(checkNameOwnership(name));
-        expect(store.getActions()).toInclude(expectedAction);
-      });
-
-      it('should store CHECK_NAME_OWNERSHIP_ERROR action', async() => {
-        await store.dispatch(checkNameOwnership(name));
-        expect(store.getActions()).toHaveActionsMatching((action) => {
-          return action.type === ActionTypes.CHECK_NAME_OWNERSHIP_ERROR &&
-            action.payload.name === name;
-        });
-      });
-
     });
   });
 });
